# specs/003-api/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Disclosure Evaluator API
  description: API for external systems to submit document evaluation requests
  version: 1.0.0
  contact:
    name: Disclosure Evaluator Team
    email: support@disclosure-evaluator.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: https://api.disclosure-evaluator.com/v1
    description: Production server
  - url: https://staging-api.disclosure-evaluator.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Development server

security:
  - ApiKeyAuth: []
  - BearerAuth: []

paths:
  /health:
    get:
      summary: Health check endpoint
      description: Check API health and availability
      tags:
        - Health
      security: []
      responses:
        "200":
          description: API is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
        "503":
          description: API is unhealthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /evaluation:
    post:
      summary: Evaluate single document
      description: Submit a single document for disclosure evaluation
      tags:
        - Evaluation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluationRequest"
            examples:
              basic_request:
                summary: Basic evaluation request
                value:
                  document_content: "This document contains personal information about John Doe."
                  context: "Request for information disclosure"
      responses:
        "200":
          description: Evaluation completed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluationResponse"
        "202":
          description: Evaluation accepted for processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AsyncResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /batch:
    post:
      summary: Submit batch evaluation
      description: Submit multiple documents for batch evaluation
      tags:
        - Batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchRequest"
            examples:
              batch_request:
                summary: Batch evaluation request
                value:
                  documents:
                    - document_id: "doc_001"
                      content: "First document content"
                    - document_id: "doc_002"
                      content: "Second document content"
                  batch_options:
                    max_concurrent: 5
                    timeout_per_document: 300
      responses:
        "202":
          description: Batch evaluation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchResponse"
        "400":
          description: Invalid batch request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /batch/{batch_id}:
    get:
      summary: Get batch status
      description: Check the status of a batch evaluation
      tags:
        - Batch
      parameters:
        - name: batch_id
          in: path
          required: true
          description: Batch identifier
          schema:
            type: string
            example: "batch_20250105_123456"
      responses:
        "200":
          description: Batch status retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchStatusResponse"
        "404":
          description: Batch not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /batch/{batch_id}/results:
    get:
      summary: Get batch results
      description: Retrieve results for a completed batch evaluation
      tags:
        - Batch
      parameters:
        - name: batch_id
          in: path
          required: true
          description: Batch identifier
          schema:
            type: string
            example: "batch_20250105_123456"
      responses:
        "200":
          description: Batch results retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchResultsResponse"
        "202":
          description: Batch still processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchStatusResponse"
        "404":
          description: Batch not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /status/{request_id}:
    get:
      summary: Get request status
      description: Check the status of a specific request
      tags:
        - Status
      parameters:
        - name: request_id
          in: path
          required: true
          description: Request identifier
          schema:
            type: string
            example: "req_20250105_123456"
      responses:
        "200":
          description: Request status retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"
        "404":
          description: Request not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication

  schemas:
    EvaluationRequest:
      type: object
      required:
        - document_content
      properties:
        document_content:
          type: string
          description: The text content to evaluate
          maxLength: 1048576
          example: "This document contains personal information about John Doe."
        context:
          type: string
          description: Additional context information
          maxLength: 512000
          example: "Request for information disclosure under FOI Act"
        output_text:
          type: string
          description: Desired output text format
          maxLength: 1024
          example: "Please provide a summary of the disclosure decision"
        provider:
          type: string
          description: LLM provider preference
          enum: [openai, anthropic, bedrock, bedrock_nova]
          example: "anthropic"
        options:
          $ref: "#/components/schemas/EvaluationOptions"

    EvaluationResponse:
      type: object
      required:
        - evaluation_id
        - document_summary
        - criteria_results
        - overall_score
        - recommendation
        - processing_time
      properties:
        evaluation_id:
          type: string
          description: Unique identifier for this evaluation
          example: "eval_20250105_123456"
        document_summary:
          type: string
          description: Brief summary of the document
          example: "Document contains personal information about an individual"
        criteria_results:
          type: array
          description: Results for each evaluation criterion
          items:
            $ref: "#/components/schemas/CriterionResult"
        overall_score:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: Overall disclosure score
          example: 0.7
        recommendation:
          type: string
          enum:
            [
              DISCLOSE,
              LIKELY_DISCLOSE,
              UNCLEAR,
              LIKELY_NON_DISCLOSE,
              NON_DISCLOSE,
            ]
          description: Final disclosure recommendation
          example: "LIKELY_DISCLOSE"
        processing_time:
          type: number
          format: float
          description: Processing time in seconds
          example: 15.3
        metadata:
          $ref: "#/components/schemas/EvaluationMetadata"

    BatchRequest:
      type: object
      required:
        - documents
      properties:
        documents:
          type: array
          description: List of documents to evaluate
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/BatchDocument"
        batch_options:
          $ref: "#/components/schemas/BatchOptions"
        callback_url:
          type: string
          format: uri
          description: Webhook URL for completion notification
          example: "https://example.com/webhook"
        priority:
          type: string
          enum: [LOW, NORMAL, HIGH, URGENT]
          default: NORMAL
          description: Processing priority
          example: "NORMAL"

    BatchResponse:
      type: object
      required:
        - batch_id
        - status
        - total_documents
      properties:
        batch_id:
          type: string
          description: Unique identifier for the batch
          example: "batch_20250105_123456"
        status:
          type: string
          enum: [PENDING, PROCESSING, COMPLETED, PARTIALLY_COMPLETED, FAILED]
          description: Current batch status
          example: "PENDING"
        total_documents:
          type: integer
          minimum: 1
          maximum: 100
          description: Total number of documents in batch
          example: 5
        processed_documents:
          type: integer
          minimum: 0
          description: Number of documents processed
          example: 0
        failed_documents:
          type: integer
          minimum: 0
          description: Number of documents that failed
          example: 0
        estimated_completion:
          type: string
          format: date-time
          description: Estimated completion time
          example: "2025-10-05T12:30:00Z"
        results:
          type: array
          description: Individual evaluation results (if completed)
          items:
            $ref: "#/components/schemas/EvaluationResponse"

    BatchDocument:
      type: object
      required:
        - document_id
        - content
      properties:
        document_id:
          type: string
          description: Unique identifier for this document
          example: "doc_001"
        content:
          type: string
          description: Document content
          maxLength: 1048576
          example: "Document content here..."
        context:
          type: string
          description: Additional context
          maxLength: 512000
          example: "Context information"
        metadata:
          type: object
          description: Document metadata
          additionalProperties: true

    EvaluationOptions:
      type: object
      properties:
        parallel_processing:
          type: boolean
          default: true
          description: Enable parallel evaluation
        max_workers:
          type: integer
          minimum: 1
          maximum: 30
          default: 5
          description: Maximum parallel workers
        timeout:
          type: integer
          minimum: 30
          maximum: 1800
          default: 300
          description: Processing timeout in seconds
        retry_attempts:
          type: integer
          minimum: 0
          maximum: 3
          default: 1
          description: Number of retry attempts

    BatchOptions:
      type: object
      properties:
        max_concurrent:
          type: integer
          minimum: 1
          maximum: 10
          default: 5
          description: Maximum concurrent evaluations
        timeout_per_document:
          type: integer
          minimum: 30
          maximum: 1800
          default: 300
          description: Timeout per document in seconds
        retry_failed:
          type: boolean
          default: true
          description: Whether to retry failed documents
        notification_webhook:
          type: string
          format: uri
          description: Webhook for notifications

    CriterionResult:
      type: object
      required:
        - criterion_id
        - criterion_name
        - score
        - reasoning
      properties:
        criterion_id:
          type: string
          description: Criterion identifier
          example: "personal_info_protection"
        criterion_name:
          type: string
          description: Human-readable criterion name
          example: "Personal Information Protection"
        score:
          type: number
          format: float
          minimum: 1.0
          maximum: 5.0
          description: Score for this criterion
          example: 3.5
        reasoning:
          type: string
          description: Detailed reasoning for the score
          example: "The document contains personal information that could identify an individual."
        steps:
          type: array
          description: Individual evaluation steps
          items:
            $ref: "#/components/schemas/EvaluationStep"

    EvaluationStep:
      type: object
      required:
        - step_number
        - step_name
        - result
        - reasoning
      properties:
        step_number:
          type: integer
          minimum: 1
          description: Step number in the evaluation process
          example: 1
        step_name:
          type: string
          description: Name of the evaluation step
          example: "Identify personal information"
        result:
          type: string
          enum: [YES, NO, UNCLEAR]
          description: Result of this step
          example: "YES"
        reasoning:
          type: string
          description: Reasoning for the step result
          example: "The document contains a full name and address."

    EvaluationMetadata:
      type: object
      properties:
        api_version:
          type: string
          description: API version used
          example: "1.0.0"
        processing_time:
          type: number
          format: float
          description: Server processing time in seconds
          example: 2.1
        rate_limit_remaining:
          type: integer
          description: Remaining rate limit
          example: 95
        rate_limit_reset:
          type: string
          format: date-time
          description: Rate limit reset time
          example: "2025-10-05T13:00:00Z"
        correlation_id:
          type: string
          description: Request correlation ID
          example: "api_20250105_123456"

    HealthResponse:
      type: object
      required:
        - status
        - timestamp
      properties:
        status:
          type: string
          enum: [healthy, unhealthy]
          description: API health status
          example: "healthy"
        timestamp:
          type: string
          format: date-time
          description: Health check timestamp
          example: "2025-10-05T12:00:00Z"
        version:
          type: string
          description: API version
          example: "1.0.0"
        uptime:
          type: number
          format: float
          description: API uptime in seconds
          example: 86400.0

    StatusResponse:
      type: object
      required:
        - request_id
        - status
        - progress
        - message
        - last_updated
      properties:
        request_id:
          type: string
          description: Request identifier
          example: "req_20250105_123456"
        status:
          type: string
          enum: [QUEUED, PROCESSING, COMPLETED, FAILED, CANCELLED]
          description: Current processing status
          example: "PROCESSING"
        progress:
          type: number
          format: float
          minimum: 0.0
          maximum: 100.0
          description: Progress percentage
          example: 75.5
        message:
          type: string
          description: Human-readable status message
          example: "Processing document 3 of 5"
        last_updated:
          type: string
          format: date-time
          description: Last status update time
          example: "2025-10-05T12:15:00Z"
        estimated_completion:
          type: string
          format: date-time
          description: Estimated completion time
          example: "2025-10-05T12:20:00Z"
        errors:
          type: array
          description: Any processing errors
          items:
            $ref: "#/components/schemas/APIError"

    BatchStatusResponse:
      type: object
      required:
        - batch_id
        - status
        - total_documents
        - processed_documents
        - failed_documents
      properties:
        batch_id:
          type: string
          description: Batch identifier
          example: "batch_20250105_123456"
        status:
          type: string
          enum: [PENDING, PROCESSING, COMPLETED, PARTIALLY_COMPLETED, FAILED]
          description: Current batch status
          example: "PROCESSING"
        total_documents:
          type: integer
          description: Total documents in batch
          example: 10
        processed_documents:
          type: integer
          description: Documents processed successfully
          example: 7
        failed_documents:
          type: integer
          description: Documents that failed
          example: 1
        progress:
          type: number
          format: float
          minimum: 0.0
          maximum: 100.0
          description: Overall progress percentage
          example: 70.0
        estimated_completion:
          type: string
          format: date-time
          description: Estimated completion time
          example: "2025-10-05T12:30:00Z"
        last_updated:
          type: string
          format: date-time
          description: Last status update time
          example: "2025-10-05T12:15:00Z"

    BatchResultsResponse:
      type: object
      required:
        - batch_id
        - status
        - results
      properties:
        batch_id:
          type: string
          description: Batch identifier
          example: "batch_20250105_123456"
        status:
          type: string
          enum: [COMPLETED, PARTIALLY_COMPLETED]
          description: Batch completion status
          example: "COMPLETED"
        results:
          type: array
          description: Individual evaluation results
          items:
            $ref: "#/components/schemas/EvaluationResponse"
        summary:
          type: object
          description: Batch processing summary
          properties:
            total_documents:
              type: integer
              description: Total documents processed
              example: 10
            successful_documents:
              type: integer
              description: Successfully processed documents
              example: 9
            failed_documents:
              type: integer
              description: Failed documents
              example: 1
            total_processing_time:
              type: number
              format: float
              description: Total processing time in seconds
              example: 45.2

    AsyncResponse:
      type: object
      required:
        - request_id
        - status
        - message
      properties:
        request_id:
          type: string
          description: Request identifier
          example: "req_20250105_123456"
        status:
          type: string
          enum: [ACCEPTED, PROCESSING]
          description: Request status
          example: "ACCEPTED"
        message:
          type: string
          description: Status message
          example: "Request accepted for processing"
        estimated_completion:
          type: string
          format: date-time
          description: Estimated completion time
          example: "2025-10-05T12:05:00Z"

    ErrorResponse:
      type: object
      required:
        - error_code
        - error_message
        - timestamp
      properties:
        error_code:
          type: string
          description: Machine-readable error code
          example: "VALIDATION_ERROR"
        error_message:
          type: string
          description: Human-readable error message
          example: "Invalid request format"
        error_type:
          type: string
          enum:
            [
              VALIDATION_ERROR,
              AUTHENTICATION_ERROR,
              AUTHORIZATION_ERROR,
              RATE_LIMIT_ERROR,
              PROCESSING_ERROR,
              TIMEOUT_ERROR,
              SERVICE_ERROR,
            ]
          description: Error category
          example: "VALIDATION_ERROR"
        correlation_id:
          type: string
          description: Request correlation ID
          example: "api_20250105_123456"
        timestamp:
          type: string
          format: date-time
          description: Error occurrence time
          example: "2025-10-05T12:00:00Z"
        details:
          type: object
          description: Additional error context
          additionalProperties: true
        resolution_guidance:
          type: string
          description: Suggested resolution steps
          example: "Please check the request format and try again"

    APIError:
      type: object
      required:
        - error_code
        - error_message
        - error_type
      properties:
        error_code:
          type: string
          description: Machine-readable error code
          example: "PROCESSING_ERROR"
        error_message:
          type: string
          description: Human-readable error message
          example: "Failed to process document"
        error_type:
          type: string
          enum:
            [
              VALIDATION_ERROR,
              AUTHENTICATION_ERROR,
              AUTHORIZATION_ERROR,
              RATE_LIMIT_ERROR,
              PROCESSING_ERROR,
              TIMEOUT_ERROR,
              SERVICE_ERROR,
            ]
          description: Error category
          example: "PROCESSING_ERROR"
        correlation_id:
          type: string
          description: Request correlation ID
          example: "api_20250105_123456"
        timestamp:
          type: string
          format: date-time
          description: Error occurrence time
          example: "2025-10-05T12:00:00Z"
        details:
          type: object
          description: Additional error context
          additionalProperties: true
        resolution_guidance:
          type: string
          description: Suggested resolution steps
          example: "Please retry the request or contact support"

tags:
  - name: Health
    description: Health check endpoints
  - name: Evaluation
    description: Single document evaluation endpoints
  - name: Batch
    description: Batch evaluation endpoints
  - name: Status
    description: Request status checking endpoints